Question_repo/requirements.txt
Django==1.11.20
django-ckeditor==5.7.0
django-js-asset==1.2.2
django-redis==4.10.0
easy-thumbnails==2.6
Pillow==6.0.0
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
xlrd==1.2.0
